+++ packages/lib/src/index.ts
@@ -2,3 +2,4 @@ export * from './env';
 export * from './utils';
 export * from './jwt';
 export * from './oauth';
+export * from './sm2';